backend/app/schemas.py
from pydantic import BaseModel
from datetime import datetime

class MessageCreate(BaseModel):
    content: str

class MessageResponse(BaseModel):
    id: int
    content: str
    response: str
    created_at: datetime

    class Config:
        orm_mode = True