backend/pyproject.toml
[tool.poetry]
name = "chatgpt-mini-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
sqlalchemy = {extras = ["asyncio"], version = "^1.4.23"}
asyncpg = "^0.27.0"
python-dotenv = "^0.19.0"
openai = "^0.27.0"